// packages/leak-sdk/src/client/types.gen.ts
export type GetCoinData = {
  body?: never;
  path?: never;
  query: {
    address: string;
    chain?: number;
  };
  url: "/coin";
};
// ...
export type GetCoinResponse = GetCoinResponses[keyof GetCoinResponses];
// ...
export type GetCoinHoldersData = {
  body?: never;
  path?: never;
  query: {
    chainId: number;
    address: string;
    after?: string;
    count?: number;
  };
  url: "/coinHolders";
};

export type GetCoinHoldersErrors = {
  /**
   * Bad request
   */
  400: unknown;
  /**
   * Internal server error
   */
  500: unknown;
};

export type GetCoinHoldersResponses = {
  /**
   * Successful operation
   */
  200: {
    zora20Token?: {
      tokenBalances: {
        pageInfo: {
          /**
           * When paginating forwards, the cursor to continue.
           */
          endCursor?: string;
          /**
           * When paginating forwards, are there more items?
           */
          hasNextPage: boolean;
        };
        /**
         * The `Int` scalar type represents non-fractional signed whole numeric values. Int can represent values between -(2^31) and 2^31 - 1.
         */
        count: number;
        edges: Array<{
          node: {
            /**
             * The `String` scalar type represents textual data, represented as UTF-8 character sequences. The String type is most often used by GraphQL to represent free-form human-readable text.
             */
            balance: string;
            ownerAddress: string;
            ownerProfile: {
              /**
               * The Globally Unique ID of this object
               */
              id: string;
              /**
               * Manually set username, or truncated wallet address if the profile isn't a GraphQLAccountProfile. For full wallet address, use the profile_id field instead.
               */
              handle: string;
              /**
               * The `Boolean` scalar type represents `true` or `false`.
               */
              platformBlocked: boolean;
              avatar?: {
                previewImage: {
                  /**
                   * The `String` scalar type represents textual data, represented as UTF-8 character sequences. The String type is most often used by GraphQL to represent free-form human-readable text.
                   */
                  blurhash?: string;
                  /**
                   * The `String` scalar type represents textual data, represented as UTF-8 character sequences. The String type is most often used by GraphQL to represent free-form human-readable text.
                   */
                  medium: string;
                  /**
                   * The `String` scalar type represents textual data, represented as UTF-8 character sequences. The String type is most often used by GraphQL to represent free-form human-readable text.
                   */
                  small: string;
                };
              };
            };
          };
        }>;
      };
    };
  };
};

export type GetCoinHoldersResponse =
  GetCoinHoldersResponses[keyof GetCoinHoldersResponses];
// ...
export type GetCoinPriceHistoryData = {
  body?: never;
  path?: never;
  query: {
    address: string;
    chain?: number;
  };
  url: "/coinPriceHistory";
};

export type GetCoinPriceHistoryErrors = {
  /**
   * Bad request
   */
  400: unknown;
  /**
   * Internal server error
   */
  500: unknown;
};

export type GetCoinPriceHistoryResponses = {
  /**
   * Successful operation
   */
  200: {
    zora20Token?: {
      oneHour: Array<{
        /**
         * Date with time (isoformat)
         */
        timestamp: string;
        /**
         * The `String` scalar type represents textual data, represented as UTF-8 character sequences. The String type is most often used by GraphQL to represent free-form human-readable text.
         */
        closePrice: string;
      }>;
      oneDay: Array<{
        /**
         * Date with time (isoformat)
         */
        timestamp: string;
        /**
         * The `String` scalar type represents textual data, represented as UTF-8 character sequences. The String type is most often used by GraphQL to represent free-form human-readable text.
         */
        closePrice: string;
      }>;
      oneWeek: Array<{
        /**
         * Date with time (isoformat)
         */
        timestamp: string;
        /**
         * The `String` scalar type represents textual data, represented as UTF-8 character sequences. The String type is most often used by GraphQL to represent free-form human-readable text.
         */
        closePrice: string;
      }>;
      oneMonth: Array<{
        /**
         * Date with time (isoformat)
         */
        timestamp: string;
        /**
         * The `String` scalar type represents textual data, represented as UTF-8 character sequences. The String type is most often used by GraphQL to represent free-form human-readable text.
         */
        closePrice: string;
      }>;
      all: Array<{
        /**
         * Date with time (isoformat)
         */
        timestamp: string;
        /**
         * The `String` scalar type represents textual data, represented as UTF-8 character sequences. The String type is most often used by GraphQL to represent free-form human-readable text.
         */
        closePrice: string;
      }>;
    };
  };
};

export type GetCoinPriceHistoryResponse =
  GetCoinPriceHistoryResponses[keyof GetCoinPriceHistoryResponses];

export type GetCoinSwapsData = {
  body?: never;
  path?: never;
  query: {
    address: string;
    chain?: number;
    after?: string;
    first?: number;
  };
  url: "/coinSwaps";
};
// ...
export type GetCoinSwapsResponse =
  GetCoinSwapsResponses[keyof GetCoinSwapsResponses];

export type GetCoinsData = {
  body?: never;
  path?: never;
  query: {
    coins: Array<{
      /**
       * The `Int` scalar type represents non-fractional signed whole numeric values. Int can represent values between -(2^31) and 2^31 - 1.
       */
      chainId: number;
      collectionAddress: string;
    }>;
  };
  url: "/coins";
};
// ...
export type GetCoinsResponse = GetCoinsResponses[keyof GetCoinsResponses];
// ...
export type GetContentCoinPoolConfigData = {
  body?: never;
  path?: never;
  query: {
    creatorIdentifier?: string;
    currencyType:
      "ETH" | "ZORA" | "CREATOR_COIN" | "CREATOR_COIN_OR_ZORA" | "CUSTOM_COIN";
  };
  url: "/contentCoinPoolConfig";
};
// ...
export type GetContentCoinPoolConfigResponse =
  GetContentCoinPoolConfigResponses[keyof GetContentCoinPoolConfigResponses];
// ...
export type GetProfileData = {
  body?: never;
  path?: never;
  query: {
    identifier: string;
  };
  url: "/profile";
};
// ...
export type GetProfileResponse = GetProfileResponses[keyof GetProfileResponses];

export type GetProfileBalancesData = {
  body?: never;
  path?: never;
  query: {
    identifier: string;
    count?: number;
    after?: string;
    sortOption?:
      | "BALANCE"
      | "MARKET_CAP"
      | "USD_VALUE"
      | "PRICE_CHANGE"
      | "MARKET_VALUE_USD";
    excludeHidden?: boolean;
    chainIds?: Array<number>;
  };
  url: "/profileBalances";
};
// ...
export type GetProfileBalancesResponse =
  GetProfileBalancesResponses[keyof GetProfileBalancesResponses];
// ...
export type GetProfileCoinsData = {
  body?: never;
  path?: never;
  query: {
    identifier: string;
    count?: number;
    after?: string;
    chainIds?: Array<number>;
    platformReferrerAddress?: Array<string>;
  };
  url: "/profileCoins";
};
// ...
export type GetProfileCoinsResponse =
  GetProfileCoinsResponses[keyof GetProfileCoinsResponses];
// ...
export type GetTokenInfoData = {
  body?: never;
  path?: never;
  query: {
    address: string;
    chainId?: number;
  };
  url: "/tokenInfo";
};

export type GetTokenInfoErrors = {
  /**
   * Bad request
   */
  400: unknown;
  /**
   * Internal server error
   */
  500: unknown;
};

export type GetTokenInfoResponses = {
  /**
   * Successful operation
   */
  200: {
    erc20Token?: {
      currency?: {
        /**
         * The `String` scalar type represents textual data, represented as UTF-8 character sequences. The String type is most often used by GraphQL to represent free-form human-readable text.
         */
        priceUsd?: string;
        /**
         * The `Int` scalar type represents non-fractional signed whole numeric values. Int can represent values between -(2^31) and 2^31 - 1.
         */
        decimals?: number;
        /**
         * The `String` scalar type represents textual data, represented as UTF-8 character sequences. The String type is most often used by GraphQL to represent free-form human-readable text.
         */
        name?: string;
        /**
         * The `String` scalar type represents textual data, represented as UTF-8 character sequences. The String type is most often used by GraphQL to represent free-form human-readable text.
         */
        symbol?: string;
        /**
         * The `String` scalar type represents textual data, represented as UTF-8 character sequences. The String type is most often used by GraphQL to represent free-form human-readable text.
         */
        icon?: string;
      };
    };
  };
};

export type GetTokenInfoResponse =
  GetTokenInfoResponses[keyof GetTokenInfoResponses];
// ...
export type GetTraderLeaderboardData = {
  body?: never;
  path?: never;
  query?: {
    week?: number;
    year?: number;
    first?: number;
    after?: string;
  };
  url: "/traderLeaderboard";
};

export type GetTraderLeaderboardErrors = {
  /**
   * Bad request
   */
  400: unknown;
  /**
   * Internal server error
   */
  500: unknown;
};

export type GetTraderLeaderboardResponses = {
  /**
   * Successful operation
   */
  200: {
    exploreTraderLeaderboard: {
      /**
       * The `Int` scalar type represents non-fractional signed whole numeric values. Int can represent values between -(2^31) and 2^31 - 1.
       */
      count: number;
      edges: Array<{
        node: {
          /**
           * The `String` scalar type represents textual data, represented as UTF-8 character sequences. The String type is most often used by GraphQL to represent free-form human-readable text.
           */
          entityName: string;
          /**
           * The `Int` scalar type represents non-fractional signed whole numeric values. Int can represent values between -(2^31) and 2^31 - 1.
           */
          score: number;
          /**
           * The `Float` scalar type represents signed double-precision fractional values as specified by [IEEE 754](https://en.wikipedia.org/wiki/IEEE_floating_point).
           */
          weekVolumeUsd: number;
          /**
           * The `Int` scalar type represents non-fractional signed whole numeric values. Int can represent values between -(2^31) and 2^31 - 1.
           */
          weekTradesCount: number;
          /**
           * The `Float` scalar type represents signed double-precision fractional values as specified by [IEEE 754](https://en.wikipedia.org/wiki/IEEE_floating_point).
           */
          weekGrossVolumeZora: number;
          traderProfile?: {
            /**
             * Manually set username, or truncated wallet address if the profile isn't a GraphQLAccountProfile. For full wallet address, use the profile_id field instead.
             */
            handle: string;
            /**
             * The Globally Unique ID of this object
             */
            id: string;
          };
        };
      }>;
      pageInfo: {
        /**
         * When paginating forwards, are there more items?
         */
        hasNextPage: boolean;
        /**
         * When paginating forwards, the cursor to continue.
         */
        endCursor?: string;
      };
    };
  };
};

export type GetTraderLeaderboardResponse =
  GetTraderLeaderboardResponses[keyof GetTraderLeaderboardResponses];
// ...
export type GetWalletTradeActivityData = {
  body?: never;
  path?: never;
  query: {
    identifier: string;
    after?: string;
    first?: number;
  };
  url: "/walletTradeActivity";
};
// ...
export type GetWalletTradeActivityResponse =
  GetWalletTradeActivityResponses[keyof GetWalletTradeActivityResponses];